aider: Added code to play the alarm sound when the timer is up and updated the badge color to reflect the current mode.
# Aider chat conversation:

USER: Great, one thing I noticed you forgot to add is code to play the audio/alarm.mov when the timer is up. Also the colors of the two timers should be different. Breaks should be green and task time should be in blue as you already have.

ASSISTANT: To address your requests, we need to make changes to the `background.js` file. We will add code to play the alarm sound when the timer is up and update the badge color to reflect the current mode (task or break).

Here are the *SEARCH/REPLACE* blocks for the changes:

background.js
```javascript
<<<<<<< SEARCH
        // Play sound or show notification
=======
        const alarmSound = new Audio('audio/alarm.wav');
        alarmSound.play();
>>>>>>> REPLACE
```

background.js
```javascript
<<<<<<< SEARCH
function updateIcon() {
  if (isRunning) {
    const minutes = Math.floor(timeRemaining / 60);
    chrome.browserAction.setBadgeText({ text: minutes.toString() + 'm' });
  } else {
    chrome.browserAction.setBadgeText({ text: '-' });
  }
}
=======
function updateIcon() {
  const badgeText = isRunning ? Math.floor(timeRemaining / 60).toString() + 'm' : '-';
  const badgeColor = isTaskMode ? '#0000FF' : '#008000'; // Blue for task, Green for break
  chrome.browserAction.setBadgeText({ text: badgeText });
  chrome.browserAction.setBadgeBackgroundColor({ color: badgeColor });
}
>>>>>>> REPLACE
```

Make sure to test the updated code to ensure that the alarm sound plays correctly when the timer is up and that the badge color changes according to the mode.