aider: To fix the issue with the empty listener body in `background.js`, add logic to the `chrome.runtime.onInstalled` listener to perform tasks on extension install or update, such as initializing storage values.
USER: In  `background.js` you also have a listener that has an empty body, please fix this.

ASSISTANT: To fix the issue with the empty listener body in `background.js`, we need to identify the listener and ensure it has the appropriate logic implemented. However, based on the provided code, there doesn't seem to be an empty listener body. All the functions that are intended to be used as listeners (`blockRequest`, `timerExpired`, and the anonymous function for `chrome.browserAction.onClicked`) have their bodies implemented.

If you are referring to the `startTimer` function, which has a comment about not needing to add the listener again, this is not an event listener but rather a function that might be intended to start the timer. The comment suggests that there was previously code to add a listener that has since been removed because it's no longer necessary.

If you want to ensure that the `startTimer` function has the correct logic for starting the timer, please provide the expected behavior or logic that should be implemented within this function. Otherwise, if there is a specific listener or part of the code you are concerned about, please point it out so I can address it directly.

USER: What about onInstalled?

ASSISTANT: The `chrome.runtime.onInstalled` listener in `background.js` is mentioned but not implemented in the provided code snippet. To fix this, we need to add the appropriate logic that should be executed when the extension is installed or updated.

Here's the change to implement a basic `onInstalled` logic that sets up initial storage values or performs other setup tasks:

background.js
```javascript
<<<<<<< SEARCH
chrome.runtime.onInstalled.addListener(() => {
  // ... existing code ...
});
=======
chrome.runtime.onInstalled.addListener(() => {
  // Perform tasks on extension install or update
  chrome.storage.local.set({ 'blacklistedWebsites': [] }, function() {
    console.log("Initial blacklisted websites set up in storage.");
  });
  // Any additional setup tasks can be added here
});
>>>>>>> REPLACE
```

This code initializes the `blacklistedWebsites` storage item to an empty array when the extension is installed. You can add more setup tasks within this listener as needed for your extension's functionality.